requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
groq>=0.11.0
playwright>=1.48.0
# After pip install, run: playwright install chromium
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
reportlab>=4.0.0
